Internal: isolate Blockifier compatibility from `SharedState`
Problem: the `SharedState` struct currently has a `contract_addresses`
hashset that maps all the contracts in the state. We do not want to
require an exhaustive set of all contracts when using real block data
from Pathfinder.

Solution: isolate the Blockifier-related code of `SharedState` (state
diff application + type conversions) to a new `TestSharedState` struct
dedicated to the task. `SharedState` is now a base struct that aims to
be compatible with all use cases (transaction reexecution for tests,
fetch the trees from an RPC for production).
