+++ crates/starknet-os/src/starknet/business_logic/fact_state/mod.rs
@@ -1,3 +1,4 @@
 pub mod contract_class_objects;
 pub mod contract_state_objects;
 pub mod state;
+pub mod test_shared_state;
